add storie and spec test, fix template names
