+++ client/src/Components/local/organisms/setting-templates/README.md
@@ -1,0 +1,27 @@
+# Setting Forces documentation
+
+This is an example readme file for the 'forces' component.
+
+## UI Example
+
+<!-- STORY -->
+
+## Code example
+
+<!-- SOURCE -->
+
+## Props
+
+<!-- PROPS -->
+
+## Usage
+
+### ES6 Import
+```js
+import SettingTemplate from './organisms/setting-templates'
+```
+
+### CommonJS
+
+```js
+const SettingTemplate = require('./organisms/setting-templates')
