+++ client/src/Components/local/organisms/setting-templates/styles.module.scss
@@ -1,0 +1,30 @@
+.mb-20 {
+    margin-bottom: 20px;
+  }
+  
+  .row {
+    display: flex;
+    width: 100%;
+    align-items: center;
+    justify-content: space-between;
+    min-height: 52px;
+  }
+  
+  .col {
+    flex: 1;
+    cursor: pointer;
+  }
+  
+  .actions {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+  }
+  
+  .color-box {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    padding: 0 15px;
+  }
+
