+++ client/src/Components/local/organisms/setting-templates/styles.module.scss.d.ts
@@ -1,0 +1,11 @@
+// This file is automatically generated.
+// Please do not change this file!
+interface CssExports {
+  'actions': string;
+  'col': string;
+  'color-box': string;
+  'mb-20': string;
+  'row': string;
+}
+export const cssExports: CssExports;
+export default cssExports;
